change #1 - package 복사를 그대로 하여 패키지 변경 add #1 - config config bean등록 -> component scan 가능하게 설정
