--- src/main/java/com/fastcampus/javaoop/service/SortService.java
@@ -1,20 +1,0 @@
-package com.fastcampus.javaoop.service;
-
-import com.fastcampus.javaoop.logic.JavaSort;
-import com.fastcampus.javaoop.logic.Sort;
-
-import java.util.List;
-
-public class SortService {
-
-    private final Sort<String> sort;
-
-    public SortService(Sort<String> sort) {
-        this.sort = sort;
-        System.out.println("### [구현체] : " + sort.getClass().getName());
-    }
-
-    public List<String> doSort(List<String> list) {
-        return sort.sort(list);
-    }
-}
