+++ src/main/java/com/fastcampus/springpractice/config/Config.java
@@ -1,0 +1,11 @@
+package com.fastcampus.springpractice.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+
+@ComponentScan("com.fastcampus.springpractice")
+@Configuration
+public class Config {
+    // 스프링 전용 config class
+}
