+++ src/main/java/com/fastcampus/springpractice/logic/Sort.java
@@ -1,4 +1,4 @@
-package com.fastcampus.javaoop.logic;
+package com.fastcampus.springpractice.logic;
 
 import java.util.List;
