+++ src/main/java/com/fastcampus/springpractice/service/SortService.java
@@ -1,0 +1,23 @@
+package com.fastcampus.springpractice.service;
+
+import com.fastcampus.springpractice.logic.Sort;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class SortService {
+
+    private final Sort<String> sort;
+
+    // @Qualifier("bubbleSort") bubbleSort를 사용하겠다는 설정
+    public SortService(@Qualifier("bubbleSort") Sort<String> sort) {
+        this.sort = sort;
+        System.out.println("### [구현체] : " + sort.getClass().getName());
+    }
+
+    public List<String> doSort(List<String> list) {
+        return sort.sort(list);
+    }
+}
