+++ src/test/java/com/fastcampus/springpractice/MainTest.java
@@ -1,8 +1,6 @@
-package com.fastcampus.javaoop;
+package com.fastcampus.springpractice;
 
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class MainTest {
