+++ src/test/java/com/fastcampus/springpractice/logic/JavaSortTest.java
@@ -1,4 +1,4 @@
-package com.fastcampus.javaoop.logic;
+package com.fastcampus.springpractice.logic;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -7,16 +7,16 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class BubbleSortTest {
+class JavaSortTest {
 
-    @DisplayName("버블소트 - 리스트를 넣으면 정렬된 결과를 보여준다.")
+    @DisplayName("자바소트 - 리스트를 넣으면 정렬된 결과를 보여준다.")
     @Test
     void givenListWhenExecutingThenReturnSortedList() {
         // Given
-        BubbleSort<Integer> bubbleSort = new BubbleSort<>();
+        BubbleSort<Integer> javaSort = new BubbleSort<>();
 
         // When
-        List<Integer> actual = bubbleSort.sort(List.of(3,2,4,5,1));
+        List<Integer> actual = javaSort.sort(List.of(3,2,4,5,1));
 
         // then
         assertEquals(List.of(1,2,3,4,5), actual);
